Refactor: Move unsupervised regime detection to nas_tas
