+++ src/utils/ml_common/optimization/tas/data_pipeline/regime_detection.py
@@ -17,7 +17,7 @@ warnings.filterwarnings('ignore')
 
 # Import existing regime detection components
 try:
-    from ..regime_analysis.unsupervised_regime_detection import UnsupervisedRegimeDetector, RegimeDetectionConfig
+    from src.utils.nas_tas.unsupervised_regime_detection import UnsupervisedRegimeDetector, RegimeDetectionConfig
     from ..regime_analysis.regime_qualification import RegimeQualifier, RegimeQualificationConfig
     REGIME_ANALYSIS_AVAILABLE = True
 except ImportError:
